+++ src/Paradigm.Enterprise.Services.BlobStorage/Configuration/BlobStorageConfiguration.cs
@@ -1,0 +1,8 @@
+﻿namespace Paradigm.Enterprise.Services.BlobStorage.Configuration;
+
+public class BlobStorageConfiguration
+{
+    public required string StorageConnection { get; set; }
+    public int? MaxRetryAttempts { get; set; }
+    public int? RetryInitialDelayMilliseconds { get; set; }
+}